backend

+++ src/main/java/com/example/CalorieCalcu/repository/RegisterDetailsRepository.java
@@ -4,6 +4,6 @@ import com.example.CalorieCalcu.models.RegisterDetails;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface RegisterDetailsRepository extends JpaRepository< RegisterDetails,Integer> {
+public interface RegisterDetailsRepository extends JpaRepository< RegisterDetails,Long> {
     Optional<RegisterDetails> findByUserName(String userName);
 }
